fix(backend): corregir errores de compilacion en identidad y perfiles
- Usuario.getIdUsuario() declaraba Long pero el campo es Integer
  (coincide con usuarios.id_usuario y con UsuarioRepository<Usuario, Integer>),
  lo que rompia la compilacion de PublicProfileService, AuthSessionService,
  RegistrationService, LoginService, PasswordResetService y
  VerificationRecordService.
- Elimina AuthService/AuthRequest: codigo huerfano que llamaba a metodos
  inexistentes de RegisterRequest y a un constructor de AuthResponse
  desactualizado; AuthController nunca lo invoca (usa RegistrationService/
  LoginService/PasswordResetService/AuthSessionService).
- PublicProfileService.toProfile() invocaba EvaluacionRepository con una
  firma de PublicProfileResponse que ya no tiene campos de reputacion;
  se retira esa llamada muerta (el constructor de prueba
  PublicProfileServiceTest tampoco la esperaba).

+++ backend/src/main/java/com/t4kash/api/marketplace/entity/Usuario.java
@@ -41,7 +41,7 @@ public class Usuario {
     @Column(name = "id_universidad")
     private Integer idUniversidad;
 
-    public Long getIdUsuario() {
+    public Integer getIdUsuario() {
         return idUsuario;
     }
